# Do not construct intermediate sets when calculating union of multiple sets