An AddressGroup can be shared by multiple NetworkPolicies and its span
is the union of the NetworkPolicies'. The syncAddressGroup method uses
"Union" method to calculate the union set. However, the method always
create a new sets and copies the original items when traversing each
set. This performs badly when the number of the NetworkPolicies increase
and the span is big.

This patch optimizes it by using a "Merge" function to avoid above cost.
The benchmark impact is as below when there are 1000 NetworkPolicies
sharing the AddressGroup and the span is 1000 Nodes:

```
name                 old time/op    new time/op    delta
SyncAddressGroup-48     417µs ± 4%     326µs ± 4%  -21.91%  (p=0.008 n=5+5)

name                 old alloc/op   new alloc/op   delta
SyncAddressGroup-48    98.9kB ± 0%    51.1kB ± 1%  -48.27%  (p=0.008 n=5+5)

name                 old allocs/op  new allocs/op  delta
SyncAddressGroup-48     1.05k ± 0%     0.05k ± 8%  -94.78%  (p=0.008 n=5+5)
```