@tnqn 2021-03-05
Merging it as jenkins-windows-networkpolicy succeeded before I fixed the typo.
Thanks for the review @antoninbas @Dyanngg @jianjuns.